feat: 하차 예약 & 승차 및 하차 인원 파악 코드 작성
다만, 아직 승차 및 하차 인원 파악에 있어서 busId를 특정지을 수 없는 문제가 발견

+++ src/main/java/capston/busthecall/domain/dto/response/CountReservationInfo.java
@@ -10,6 +10,6 @@ import lombok.*;
 @Builder(toBuilder = true)
 public class CountReservationInfo {
 
-    private int onboard;
-    private int offboard;
+    private Long onboard;
+    private Long offboard;
 }
